refactor: Project structure (#6)
* refactor: Project structure

* fix: README import link
